Add text to auto-format query method
As auto-format rules run before text is inserted into the editor, when you use the query option to control validity you cannot access the last thing the user entered. In the case of code-blocks which should be '```` the query method can only derive '``' from the editor (currentNode).

This PR ensure the last thing typed is also passed into the query method.