Update `lodash.last` import in `getLastEntryEnclosedInToggle.ts`
Module not found: Error: Can't resolve 'lodash/last' in '/app/node_modules/@udecode/plate-toggle/dist'
Did you mean 'last.js'?
BREAKING CHANGE: The request 'lodash/last' failed to resolve only because it was resolved as fully specified
(probably because the origin is strict EcmaScript Module, e. g. a module with javascript mimetype, a '*.mjs' file, or a '*.js' file where the package.json contains '"type": "module"').
The extension in the request is mandatory for it to be fully specified.
Add the extension to the request.