:recycle: Refactor code to address suggested amends
As the `<LinkFloatingToolbar />` can just be altered on the consumer's side we have removed the ability to pass a callback to also execute on the `onKeyDownCapture` event. By doing this we make the default behaviour to `preventDefault` and allow the consumer to make the decision as to whether, or not, to implement their own callback.