updates example code imports to match instructions
Without the updated imports, the example code will not run. Including `PlateContent` and the full `Editor` import provides context about which components are belong to Plate and which are custom, helping bring clarity to the next explainer paragraph.

Also, as Plate is a "collection of re-usable components that you can copy and paste into your apps" if feels like a mistake that you cannot copy/paste the getting-started examples.